Added TradeLine patch to pass a context argument during the retreival of the TradeLine's game object.
started but have disabled a patch designed to do the same as above but for ShowInventoryActionMenu.

Added two "safe" retreival extension methods for getting blueprints that inherit from a specified blueprint.

Deleted a 63 mb folder that had a copy of the base game's "managed" folder that was being uploaded to the steam workshop.